requirements.txt
pandas
matplotlib
faiss-cpu
sentence-transformers
openai
python-dotenv
beautifulsoup4
requests
numpy
